feat: CODEF 연결, 주카드 등록 (#43)

+++ src/main/java/ewha/lux/once/domain/home/entity/Store.java
@@ -24,9 +24,9 @@ public class Store extends BaseEntity {
     private String address;
 
     @Column(name = "x")
-    private float x;
+    private double x;
 
     @Column(name = "y")
-    private float y;
+    private double y;
 
 }
